rfc4880bis: Convert rfc4880 to pandoc format.
This is a first step towards RFC4880bis.  Running

   pandoc2rfc abstract.mkd middle.mkd back.mkd

creates a draft.txt with is mostly identical to RFC4880.  The biblio
entries for the normative section are a bit different and the those
from the informational section are missing.  Tables are used instead
of simple list to allow the use of references.  Some parts are copied
verbatim from the RFC without new pandoc markup.  Author addresses and
the boilerplate information may not be correct.