web: Add header menu items also as submenu.
According to a report from Andrey Utkin posted to gnupg-users, some
people won't notice that they need to click on the top menu entry to
get to the donation page.  This patch fixes that for the important 3
top menu items.

What still needs to be fixed is that in the permanently show submenu,
the first item needs to be made passive.