web: Do not build an index.
* tools/build-website.sh: Disable VC backends.
* web/share/gpgweb.el (gpgweb-setup-project): Do not build the index.
--

The index requires an change in org-mode because our goal is to be
able to build from a read-only checkout.  Even w/o the index there is
still a problem building it because our use of org-publish-find-title
in gpgweb-postprocess-html results in an access to a killed buffer if
the checkout is read-only.

The actual plan is to build the web site from a script which delegates
the actual building to a separate user with limited permissions.
After the building the rendered files are copied to the web server
dir.  The command to run the rendering will be

  userv otheruser gpgweb-render

with the userv script for otheruser

=============================================
If ( glob calling-user myuser
     & glob service gpgweb-render
   )
       reset
       suppress-args
       cd /home/wk/s
       execute /home/myuser/bin/build-www.gnupg.org.sh --render
else
       error Nothing to do for this user and service
fi
quit
=============================================