blog: Finish Neals blog post
We started a mail discussion on this but unfortunately Neal's mail
server died during his vacation and thus we could not finish that
discussion.  Instead I applied the suggested changes by dkg and
Justus.  I also removed the misleading claim that a mail address
search for anna@examle.org would also return a key for
nanna@examle.org; the README always mentioned how to specify a user ID
and gives this example:

    * By an email address:

	"<heinrichh@uni-duesseldorf.de>"

Thus the angle brackets make sure that a complete mail match is done.
Nevertheless, the mentioned patch improves the selection a lot.

Signed-off-by: Werner Koch <wk@gnupg.org>