swdb: Rename some tags.
We now use the tag "gnupg22" instead of "gnupg21" for the 2.2 branch.
The next development branch 2.3.x will then be availabale under the
tag "gnupg24" so that there is always a fixed tag for each stable
branch.  For backward compatibility we keep the "gnupg21" for a while.

Also rename "gnupg" to "gnupg20" for the old-stable branch.  We keep
"gnupg" as a duplicate until the release of 2.2.

Adjusted all references on the website.

Signed-off-by: Werner Koch <wk@gnupg.org>