ox-gpgweb: Implement virtual headlines and HTML_CLASS property.
This also chnanges how HTML_CONTAINER_CLASS is used.  That is not a
problem here becuase we are cot yet using that property.  With these
changes the following markup is possible (example):

  * Top header
    :PROPERTIES:
    :HTML_CLASS: centered
    :END:

  ** <div class="pure-g">                    :html:virtual:

  ** Foo
     :PROPERTIES:
     :HTML_CONTAINER_CLASS: pure-u-1 pure-u-lg-1-3
     :END:

    Foo is about bar.

  ** </div><!-- end pure-g -->               :html:virtual:

The HTML_CLASS property puts a class="centered" into the <h> tag.

The HTML_CONTAINER_CLASS property encloses its <h> section with
another div:

  <div class="pure-u-1 pure-u-lg-1-3">
  <h3>Foo</h3>
  <div class="outline-text-1" id="text-1-2">
  <p>Foo is about bar.</p>
  </div>
  </div>

The "html" tag (the "virtual" is right now only informational) inserts
the given HTML directly at this place into the output.  Note that this
can't be done with #+HTML: because that would distrub the sectioning.