web: Add mail address to the imprint.
We received a letter from the Landesanstalt für Medien NRW today,
where they complained that we need to show a mail address because they
consider the "Telemedienangebot www.gnupg.org" as business oriented.
The link to the g10 code import seems not to be sufficient for them.

It is easiest to just add it and be done.

Date of letter: 2019-04-26
File reference: TM_19_107
(Handwritten signature)