web: Disable Stripe donations
We are used for card testing on large scale and thus we would first
need to install extra screenings.  Further, begiinig in September we
seem to need to get additional authentication from the users.  Thus it
is questionable whether we will re-enable credit card donations
again.  It is just too much overhead.

This removes payment options from the page and also adds an extra
check into the CGI.